lilac: Add IMS support for China Mobile, Unicom and Telecom
These carriers do not support 3g, only cdma and 4g. We don't support cdma and because of that voice calls won't work for them unless VoLTE is enabled.

Neither of them are supported in stock, however we do have the mbn for China Mobile. Let's enable support for lilac, maple_dsds and poplar_dsds.

It was reported that using the ir51_ir92_ims_gte will make IMS work for the remaining two carriers.

So, to enable IMS for them, the following modifications are done:
* Map China Unicom simid's (S285.1 – S285.3) to a new custid (928)
* Map China Telecom simid's (S8760.1 – S8760.2) to a new custid (929)
* Add a modem.conf file containing china_mobile_cn_ims for custid 318 and copy it to system
* Add a modem.conf file containing ir51_ir92_ims_gte for custid 928 and 929 and copy them to system
* Enable VoLTE (and VT in case of China Mobileas for poplar) in CarrierConfig overlays for these carriers